moves common functions
